% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMSSNE}
\alias{plotMSSNE}
\alias{plotFMSSNE}
\alias{plotMSTSNE}
\alias{plotFMSTSNE}
\title{Plot reduced dimensions}
\usage{
plotMSSNE(object, ..., ncomponents = 2)

plotFMSSNE(object, ..., ncomponents = 2)

plotFMSTSNE(object, ..., ncomponents = 2)

plotMSTSNE(object, ..., ncomponents = 2)
}
\arguments{
\item{object}{A `SingleCellExperiment` object.}

\item{...}{Additional arguments to pass to
[scater::plotReducedDim()].}

\item{ncomponents}{`numeric(1)` indicating the number of
dimensions components to plot This can also be a numeric
vector, see [scater::plotReducedDim()] for details.}
}
\value{
A `ggplot` object.
}
\description{
Wrapper fonctions to create plots for specific types
    of reduced dimension results produced by the various `fmsne`
    dimensionality reductions. The function follow the `scater`
    package's syntax.
}
\details{
As for the equivalent functions from the `scater` package, each
function is a convenient wrapper around [scater::plotReducedDim()]
that searches the [SingleCellExperiment::reducedDims()] slot for
an appropriately named dimensionality reduction result:

- "FMSTSNE" for `plotFMSTNSE()`
- "FMSSNE" for `plotFMSNSE()`
- "MSTSNE" for `plotMSTNSE()`
- "MSTSNE" for `plotMSNSE()`

Its only purpose is to streamline workflows to avoid the need to
specify the ‘dimred’ argument.
}
\seealso{
[runFMSSNE()], [runFMSSNE()], [runMSTSNE()] and [runMSSNE()] for
the functions that actually perform the calculations.

[scater::plotReducedDim()] for the underlying plotting function.
}
\author{
Laurent Gatto
}
